Chain same colored objects
This simulates chains of the atoms. Currently, the logic is simple and
only chains atoms with the same color. In future, this will be modified
so that it will reflect reactions entered by the user. The purpose of
this commit is just to test that such kind of behavior can be
implemented in UI. Splitting atoms should be tested in same way.